chore: delete KeyboardShortcut
